+++ ics/alerts.py
@@ -1,13 +1,16 @@
 import logging
 from datetime import datetime
 
-import config
 import mail
+from environment import HOSTNAME
+from environment import ICS_HOME
+from environment import ICS_CLUSTER_NAME
+from environment import ICS_ALERT_LOG
+from environment import ICS_ALERT_RECIPIENTS
 
 logger = logging.getLogger(__name__)
 
-alert_log_file = config.ICS_LOG + 'alerts.log'
-alert_html_template = config.ICS_HOME + '/ics/alert.html'
+alert_html_template = ICS_HOME + '/ics/alert.html'
 
 
 class AlertSeverity:
@@ -17,16 +20,15 @@ class AlertSeverity:
     INFO = 'info'
 
 
-
 def send_alert(resource, severity, reason='', msg=''):
 
     resource_name = resource.name
     group_name = resource.attr['Group']
-    system_name = config.CLUSTER_NAME
+    system_name = ICS_CLUSTER_NAME
     event_time = str(datetime.now().strftime("%m/%d/%Y %H:%M:%S"))
-    host_name = config.HOSTNAME
+    host_name = HOSTNAME
 
-    with open(alert_log_file, 'a+') as FILE:
+    with open(ICS_ALERT_LOG, 'a+') as FILE:
         log_str = ' '.join([event_time, severity, system_name, group_name, resource_name, reason])
         FILE.write(log_str + '\n')
 
@@ -35,8 +37,8 @@ def send_alert(resource, severity, reason='', msg=''):
         html = FILE.read()
     formatted_html = html.format(**locals())
 
-    for recipient in config.ALERT_RECIPIENTS:
-        from_addr = 'ics@' + config.HOSTNAME
+    for recipient in ICS_ALERT_RECIPIENTS:
+        from_addr = 'ics@' + HOSTNAME
         to_addr = recipient
         subject = 'ICS {} Alert - {}'.format('Warning', resource.name)
         try: