+++ ics/icsstart.py
@@ -4,12 +4,15 @@ import sys
 
 import config
 import utils
+from environment import ICS_HOME
+from environment import ICS_VERSION
+
 
 
 def start_server():
     """Start server by creating new process"""
     python_bin = sys.executable
-    cmd = [python_bin, config.ICS_HOME + '/ics/icsserver.py']
+    cmd = [python_bin, ICS_HOME + '/ics/icsserver.py']
     pid = subprocess.Popen(cmd).pid
     utils.create_pid_file(pid)
 
@@ -23,7 +26,7 @@ if __name__ == '__main__':
     args = parser.parse_args()
 
     if args.version is True:
-        print('Version: ' + config.ICS_VERSION)
+        print('Version: ' + ICS_VERSION)
     else:
         if utils.check_running():
             print('ERROR: Server is already running')